Do not try to sync steps at mysql_install_db
When a new node joins the galera cluster it is first initiated with the
mysql_install_db script. Only the new node needs to perform this so
trying to sync the pacemaker nodes at this point doesn't make sense.

Additionally the following lines waits for all nodes to be in sync
before continuing.