database: Allow bootstrap openstack dbs on a separate node
The "pg2mariadb_preparation" can be inserted into the run list of a
controller node (in the HA case only one node per controller-cluster).

It will create all databases and -users for the deployed openstack
services on an existing mariadb database (while the cloud continues to
use postgresql).

It will also create temporary override config files for all the services
for use with the respective "db_sync" tools of the services. Currently
it doesn't call the db_sync tools by itself (to be addressed in a follow
up).

FIXME:
* Avoid hardcoded name for the mariadb deployment
* Add support for missing barclamps
* call the correct db_sync tools
* cleanup files after success